Make attributes configurable via option
This makes the package more flexible. IDs are often used as anchors for
navigation on the page, so they should not get removed even if no CSS
is attached to them.

It is considered best practice to not style IDs at all and only use them
as anchor links. Therefore many people do not want ID selectors to be
filtered out.

This change makes it possible to easily exclude IDs from the list of
filtered attributes.